api: extract datetime encode/decode from class
Extract tarantool.Datetime encode and decode to external functions. This
is a breaking change, but since there is no tagged release with Datetime
yet and API was more internal rather than public, it shouldn't be an
issue.

Follows #204
